requirements.txt
openmeteo-requests
requests-cache
retry-requests

pandas
numpy
requests
python-dotenv

matplotlib
seaborn

scikit-learn
xgboost
shap

hopsworks
hsfs

fastapi
uvicorn
streamlit
